requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Gunicorn==22.0.0
psycopg2-binary==2.9.9
Pillow==10.4.0
Flask-Mail==0.10.0
python-dotenv==1.0.1
Flask-Migrate==4.0.7